Attempt to move decorated function to new module for Windows
