+++ src/vws/vws.py
@@ -6,7 +6,6 @@ import base64
 import io
 import json
 from datetime import date
-from time import sleep
 from typing import Dict, List, Optional, Union
 from urllib.parse import urljoin
 
@@ -15,6 +14,7 @@ from requests import Response
 from vws_auth_tools import authorization_header, rfc_1123_date
 
 from vws._result_codes import raise_for_result_code
+from vws._wait_for_target_processed import foobar
 from vws.reports import (
     DatabaseSummaryReport,
     TargetRecord,
@@ -253,38 +253,6 @@ class VWS:
         )
         return target_record
 
-    def _wait_for_target_processed(
-        self,
-        target_id: str,
-        seconds_between_requests: float,
-    ) -> None:
-        """
-        Wait indefinitely for a target to get past the processing stage.
-
-        Args:
-            target_id: The ID of the target to wait for.
-            seconds_between_requests: The number of seconds to wait between
-                requests made while polling the target status.
-
-        Raises:
-            ~vws.exceptions.AuthenticationFailure: The secret key is not
-                correct.
-            ~vws.exceptions.Fail: There was an error with the request. For
-                example, the given access key does not match a known database.
-            TimeoutError: The target remained in the processing stage for more
-                than five minutes.
-            ~vws.exceptions.UnknownTarget: The given target ID does not match a
-                target in the database.
-            ~vws.exceptions.RequestTimeTooSkewed: There is an error with the
-                time sent to Vuforia.
-        """
-        while True:
-            report = self.get_target_summary_report(target_id=target_id)
-            if report.status != TargetStatuses.PROCESSING:
-                return
-
-            sleep(seconds_between_requests)
-
     def wait_for_target_processed(
         self,
         target_id: str,
@@ -318,8 +286,12 @@ class VWS:
             ~vws.exceptions.RequestTimeTooSkewed: There is an error with the
                 time sent to Vuforia.
         """
-        from _wait_for_target_processed import foobar
-        foobar(vws_client=self)
+        foobar(
+            vws_client=self,
+            target_id=target_id,
+            timeout_seconds=timeout_seconds,
+            seconds_between_requests=seconds_between_requests,
+        )
 
     def list_targets(self) -> List[str]:
         """
